requirements.txt
apex==0.9.10dev
clip==0.2.0
diffusers==0.32.2
einops==0.8.2
fairscale==0.4.13
flash_attn==2.8.3
huggingface_hub==0.33.1
matplotlib==3.10.8
numpy==2.4.2
omegaconf==2.3.0
Pillow==12.1.1
pytorch_lightning==2.5.2
PyYAML==6.0.2
Requests==2.32.5
scikit_learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
torch==2.6.0
torch_fidelity==0.4.0b0
torchmetrics==1.7.3
torchvision==0.21.0
tqdm==4.67.1
transformers==4.53.0
webdataset==1.0.2
